\name{call.npci}
\alias{call.npci}
\title{
the s4 class function
}
\description{
the s4 class function}
\usage{
call.npci(.Object)
}
\arguments{
  \item{.Object}{the fCI object}
}
\details{
  The S4 method will compute DEGs and save the results 
  to the original s4 object .Object
}
\value{
  \item{NA }{No values will be returned}
}
\references{http://software.steenlab.org/fCI/}
\author{Shaojun Tang}
\note{TBD}

\seealso{TBD}
\examples{
print("See README")
}
\keyword{ ~kwd1 }
