\name{fci.data}
\docType{data}
\alias{fci.data}
\title{data frame of gene expression}
\description{
  This data set gives the gene expression values for 
  multiple control and case samples.
}
\usage{fci.data}
\format{a matrix containing 1043 genes and 4 samples.}
\source{software.steen.org}
\value{
 \item{dataframe}{A data frame of expression values}
}
\references{http://software.steenlab.org/fCI/}
\keyword{datasets}