\name{figures}
\alias{figures}
\title{
generic function to draw figures of the current analysis}
\description{
generic function to draw figures of the current analysis}
\usage{
figures(.Object)
}
\arguments{
  \item{.Object}{
}
}
\details{
TBD
}
\value{
  \item{NA }{No values will be returned}
}
\references{http://software.steenlab.org/fCI/}
\author{Shaojun Tang}
\note{TBD}

\seealso{TBD}
\examples{
print("See README")
}
\keyword{ ~kwd1 }
