\name{get.fold.large.step}
\alias{get.fold.large.step}
\title{generate fold change cutoff values for fCI divergence computation}
\description{
generate fold change cutoff with a large step of 0.5 fold}
\usage{
get.fold.large.step()
}
\details{TBD
}
\value{
 \item{ fold_values }{A vector of predefined fold changes for fCI computation}

}
\references{http://software.steenlab.org/fCI/}
\author{Shaojun Tang}
\note{TBD}

\seealso{TBD}
\examples{
get.fold.large.step()
}
\keyword{ ~kwd1 }
