\name{get.protein.fold.step}
\alias{get.protein.fold.step}
\title{generate fold-change cutoff on proteomics data (with large steps 
of 0.2-0.5 fold)}
\description{generate fold-change cutoff on proteomics data (with large steps 
of 0.2-0.5 fold)
}
\usage{
get.protein.fold.step()
}
\details{TBD
}
\value{
  \item{folds}{returning a vector of recommended fold ratios for proteomic 
  study}
}  
\references{http://software.steenlab.org/fCI/}
\author{Shaojun Tang}
\note{TBD}

\seealso{TBD}
\examples{
  get.protein.fold.step()
}
\keyword{ ~kwd1 }
