\name{get.rna.fold.step}
\alias{get.rna.fold.step}
\title{generate fCI fold-change cutoff values for typical RNA-Seq data}
\description{generate fCI fold-change cutoff values for typical RNA-Seq data
}
\usage{
get.rna.fold.step()
}
\details{TBD
}
\value{
  \item{folds}{a vector of fold changes fCI used for divergence computation}
}
\references{http://software.steenlab.org/fCI/}
\author{Shaojun Tang}
\note{TBD}

\seealso{TBD}
\examples{
get.rna.fold.step()
}
\keyword{ ~kwd1 }
