\name{intersect.of.lists}
\alias{intersect.of.lists}
\title{find the common values of all vectors of a list}
\description{
  find the common values of all vectors of a list
}
\usage{
intersect.of.lists(vectorlist)
}
\arguments{
  \item{vectorlist}{
  a list of list values which we want to use to find common values
  }
}
\details{TBD
}
\value{
  \item{intersects}{the common values of lists}
}
\references{http://software.steenlab.org/fCI/}
\author{Shaojun Tang}
\note{TBD}

\seealso{TBD}
\examples{
  
  print("this function will be disabled!")
  
}
\keyword{ ~kwd1 }
