\name{is.installed}
\alias{is.installed}
\title{package}
\description{
  test if a package is installed in the R library
}
\usage{
is.installed(mypkg)
}
\arguments{
  \item{mypkg}{ a R library name, such as FNN
}
}
\details{TBD
}
\value{
\item{installation}{boolean value indicating the installation}

}
\references{http://software.steenlab.org/fCI/}
\author{Shaojun Tang}
\note{TBD}

\seealso{TBD}
\examples{
  is.installed('fCI')
}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
