\name{multi.dimensional.fci.data}
\docType{data}
\alias{multi.dimensional.fci.data}
\title{data frame of gene expression}
\description{
  This data set gives the gene expression values for 
  14204 genes and the control and case samples were
  generated at two time points (bivariate data).
}
\usage{fci.data}
\format{a matrix containing 14204 genes and 8 samples.}
\source{software.steen.org}
\value{
 \item{dataframe }{A data frame of expression values}
}
\references{http://software.steenlab.org/fCI/}
\keyword{datasets}