\name{normalization}
\alias{normalization}
\title{generic function to normalize gene expression matrix
}
\description{
generic function to normalize gene expression matrix
}
\usage{
normalization(.Object)
}
\arguments{
  \item{.Object}{ the predefined class object (i.e  fCI=new("NPCI"))
}
}
\details{
TBD
}
\value{
  \item{NA }{No values will be returned}

}
\references{http://software.steenlab.org/fCI/}
\author{Shaojun Tang}
\note{TBD}

\seealso{TBD}
\examples{
print("See README")
}
\keyword{ ~kwd1 }
