\name{populate}
\alias{populate}
\title{generic function to populate the fCI object based on provided data
}
\description{
generic function to populate the fCI object based on provided data
}
\usage{
populate(.Object)
}
\arguments{
  \item{.Object}{
}
}
\details{
TBD
}
\value{
  \item{NA }{No values will be returned}

}
\references{http://software.steenlab.org/fCI/}
\author{Shaojun Tang}
\note{TBD}

\seealso{TBD}
\examples{
print("See README")
}
\keyword{ ~kwd1 }
