\name{report.target.summary}
\alias{report.target.summary}
\title{generate the results (gene ids) in the data frame 
}
\description{
generate the results (gene ids) in the data frame 
}
\usage{
report.target.summary(pairwise.diff.gene.ids)
}
\arguments{
  \item{pairwise.diff.gene.ids}{ a list of the the differentially expression 
  genes (its index)  for each pairwise fCI analysis. 
}
}
\details{
TBD
}
\value{
  \item{NA }{No values will be returned}

}
\references{http://software.steenlab.org/fCI/}
\author{Shaojun Tang}
\note{TBD}

\seealso{TBD}
\examples{
print("See README")
}
\keyword{ ~kwd1 }