\name{show.targets}
\alias{show.targets}
\title{display the gene ids that are identified to be differentially regulated
}
\description{
display the gene ids that are identified to be differentially regulated
}
\usage{
show.targets(.Object)
}
\arguments{
  \item{.Object}{ the class object, for example, fCI=new("NPCI")
}
}
\details{
TBD
}
\value{
  \item{NA }{No values will be returned}

}
\references{http://software.steenlab.org/fCI/}
\author{Shaojun Tang}
\note{TBD}

\seealso{TBD}
\examples{
print("See README")
}
\keyword{ ~kwd1 }