\name{summarize-methods}
\docType{methods}
\alias{summarize-methods}
\alias{summarize,NPCI-method}
\title{result summerization}
\description{
 summerize the result after fCI computation is done
}
\section{Methods}{
\describe{

\item{\code{signature(.Object = "NPCI")}}{
}
}}
\keyword{methods}
\keyword{ ~~ other possible keyword(s) ~~ }
