\name{two.sample.log.ratio}
\alias{two.sample.log.ratio}
\title{compute the log ratios of two vectors}
\description{compute the log ratios of two vectors
}
\usage{
two.sample.log.ratio(a, b)
}
\arguments{
  \item{a}{ a vector of numeric values (value must be greater than 0)
}
  \item{b}{ a vector of numeric values (value must be greater than 0)
}
}
\details{TBD
}
\value{
\item{ratios}{the log ratios of two vectors}
}
\references{http://software.steenlab.org/fCI/}
\author{Shaojun Tang}
\note{TBD}

\seealso{TBD}
\examples{
  
  a=10
  b=2
  two.sample.log.ratio(a, b)

}
\keyword{ ~kwd1 }
