\name{venndiagram}
\alias{venndiagram}
\title{generate a venn diagram to show the differentially expression summaries 
accross pairwise fCI analysis
}
\description{
generate a venn diagram to show the differentially expression summaries 
accross pairwise fCI analysis
}
\usage{
venndiagram(.Object)
}
\arguments{
  \item{.Object}{ the class object, i.e, fci=new("NPCI")
}
}
\details{
TBD
}
\value{
  \item{NA}{No values will be returned}

}
\references{http://software.steenlab.org/fCI/}
\author{Shaojun Tang}
\note{TBD}

\seealso{TBD}
\examples{
print("See README")
}
\keyword{ ~kwd1 }
